requirements.txt
scikit-learn==0.24.2
matplotlib==3.7.5
apache-airflow-providers-amazon==8.19.0